Retry on exceeding deadline obtaining the status of a running execution
Remote execution of actions is handled via long-running operations.
Here we have to be careful with the involved status codes: there
is the status code of the operation and the response contains a
faild that also happens to be a status code. The protocol states

  Errors discovered during creation of the `Operation` will be
  reported as gRPC Status errors, while errors that occurred while
  running the action will be reported in the `status` field of
  the `ExecuteResponse`

So we have to distinguish between two kinds of DEADLINE_EXCEEDED.
- If reported by the rpc, it means, we failed to obtain the status
  of the ongoing action in a reasonable amount of time; here we
  can do nothing but retry.
- If we obtain an answer and that answer has state DEADLINE_EXCEEDED
  this means "The execution timed out."; hence we must not retry
  and report the result properly to the user.