test serve_target_cache_hit: Remove wrong check
One of the analysis commands marked as expected to fail fails from
wrongly passing the launcher instead of the stated reasons. As the
behaviour between standalone and separate serve endpoint versions
would become inconsistent after fixing the command line, remove
this check completely. This does not negatively influence the scope
of the test script.