just-lock design: Add fields 'type','mirrors','subdir' to 'archive' source
The 'type' field is optional and informs both the way to unpack
the archive and which type imported file repositories should be
rewritten as in the output configuration. Mirrors the 'just-mr'
types, with options for tarballs and zip-like archives, defaulting
to tarballs if missing.

The 'mirrors' field is treated the same as for 'git' sources.

The 'subdir' field is optional and accounts for the fact that the
actual root of the source repository might be a subpath in the
unpacked archive, as opposed to Git repositories where it is
reasonable to expect that the sources root is the top-level
directory.