prune bootstrap sources
Originally, it was a good idea to add the full "bin" directory to
the bootstrap sources; all scripts there where used in some form
or another during bootstap and by taking the whole directory, we
wouldn't have to update the target when adding new scripts.

However, times have changed. The set of scripts neede for bootstrapping
is stable for quite a while now; on the other hand actively-developped
tools (like just-lock and just-import-git) now reside in "bin".
So, to avoid unnecessary runs of the bootstrap tests, prune the
dependency set.

While there, also mark as tainted.