target serving: do not locally log user erros as errrors
When a bad request is uploaded, log this either at Info level or,
if the kind of error is unlikely by a normal just client, at Warning
level. Errors in the local log should be reserved for those cases
where something went wrong locally, like failing to acquire a log
or creating a temporary directory, etc.