drop #ifdef __unix__ preprocessor guards
... as those do not increase portability; in fact, there are quite
a few UNIX-like systems that do not set this preprocessor macro.
Instead, document that we only support sufficiently UNIX-like systems
in the documentation and also documet where we succesfully built
and tested the tool (giving an implicit hint where we did not try).
In the code, however, just try to include the relevant files.