GitRepo: Methods expected to use a logger should do so...
...in all return paths, including in reporting caught exceptions.
In this way give the opportunity for any calling AsyncMap to
receive an expected fatal logger call on failure and thus be able
to shut down gracefully. This is in line with the AsyncMap design,
where the loggers are assumed to be safe to call by a consumer.

(cherry-picked from ae33d0f287e83769dbef9287e64006e1c0c463e9)