Add end-to-end test for serving complex targets
Add a test asking serve to provide a target where both, the dependency
as well as the action graph has both a wide node (a target with a
large number of direct dependencies) as well as a deep node (a node
where the chain of dependencies is long). That target is requested
serveral times in parallel to verify that such targets can also be
served under load.

The process actually tested is the `just serve` process implicit
to the "with serve" rule.