Invocation server: also show counts in headings
Often, the simple number, e.g., for failed actions can already give
valuable information. For example, when investigating flakiness,
the count of failed actions already gives a hint whether a change
increased or decreased flakiness which can be valuable before even
investigating if the nature of the failure has changed. As we have
that information available and an additional number in a heading
does not clutter the page too much, let's just show it.