split/splice: drop algorithm negociation prototype
So far, blob split/splice was an extension of our build tool; now
it got standardized. Therefore we can, and hence should, use the
official proto file without modifications. As algorithm negociation
is not yet standardized, drop it for now; we will readd it, once
a standard is agreed upon.