BazelResponse: Upload symlinks as explicit blobs in compatible mode
The default remote-execution protocol (which is used in compatible mode)
defines symbolic links as embedded fields within a Directory message. No
separate blobs will be stored in CAS, but they are required, e.g., when a tree
is installed. To fix that issue, we upload any symbolic link received from the
remote execution either as explicit output or as part of an output Directory
message as separate blob to the remote CAS, similar how we handle output
Directory messages themselves.