just-mr: gracefully handle empty path string
The empty string is not a valid path. So, if a "file" repository
sets the empty string as value for "path" we can either report
an error or tacitly treat is as "." as we do for module paths; in
any case we should not crash. Fix the crash by treating "" as if
it were ".".