just-mr(1): clarify the meaning of exit code 66
A missing or unknown subcommand is a syntactical error of the command
line and hence reported as 67. The only place where exit code 66 is
reported is if we could parse command line but failed to implement
that newly added subcommand. For compatibility reason, leave the
exit codes as they are, but document that 66 is an internal error
in the implementation of `just-mr`.