BlobTree{,Ptr}: add instance of std::hash
As we're using instances of std::unordered_set<BlobTree> we need
the corresponding hash instances. As hashes are allowed to have
collisions, it is sound to only use the blob for hashing. Moreover,
in the intended use case, that blob even describes the data uniquely.

This omission was discovered when updating from gcc 14.2 to gcc 14.3
together with the corresponding C++ libraries.